+++ tests/tsconfig.json
@@ -4,7 +4,8 @@
         // Reset rootDir to default to make rootDirs take effect
         "rootDir": null,
         "rootDirs": ["../components", "."],
-        "types": ["vitest/globals"]
+        "noEmit": true,
+        "emitDeclarationOnly": false
     },
     "include": ["../components", "**/*.ts"],
     "exclude": []
